requirements.txt
pandas  
python-telegram-bot